+++ src/app/models/daos/SerialDeviceDao.ts
@@ -1,4 +1,5 @@
 import { SerialDeviceEntitie } from "../domains/SerialDeviceEntitie";
+import { SerialDeviceDbError } from "../modelErrors/serialDeviceErrors";
 import { SerialDeviceInterface } from "../modelsInterfaces/SerialDeviceInterface.js";
 
 export abstract class SerialDeviceDao {
@@ -22,6 +23,8 @@ export abstract class SerialDeviceDao {
 
   public static async delete(keys: Object): Promise<void> {
     const devices = await SerialDeviceEntitie.find(keys);
+    if(!devices.length)
+      throw new SerialDeviceDbError('Não existe um SerialDevice cadastrado com esse id!');
     for (const device of devices) {
       await device.delete();
     }
